+++ src/main/java/kitchenpos/product/application/ProductService.java
@@ -1,6 +1,6 @@
 package kitchenpos.product.application;
 
-import kitchenpos.dao.ProductDao;
+import kitchenpos.product.dao.ProductDao;
 import kitchenpos.product.domain.Product;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
