+++ src/main/java/kitchenpos/product/ui/ProductRestController.java
@@ -2,8 +2,8 @@ package kitchenpos.product.ui;
 
 import kitchenpos.product.application.ProductService;
 import kitchenpos.product.domain.Product;
-import kitchenpos.request.ProductCreateRequest;
-import kitchenpos.response.ProductResponse;
+import kitchenpos.product.request.ProductCreateRequest;
+import kitchenpos.product.response.ProductResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
